Moved all section logic into result
https://coveord.atlassian.net/browse/KIT-1147
