+++ packages/atomic/src/components/atomic-result/atomic-result.tsx
@@ -7,6 +7,17 @@ import {
   getResultDisplayClasses,
 } from './atomic-result-display-options';
 
+const resultSectionTags = [
+  'atomic-result-section-visual',
+  'atomic-result-section-badges',
+  'atomic-result-section-actions',
+  'atomic-result-section-title',
+  'atomic-result-section-title-metadata',
+  'atomic-result-section-emphasized',
+  'atomic-result-section-excerpt',
+  'atomic-result-section-bottom-metadata',
+] as const;
+
 /**
  * The `atomic-result` component is used internally by the `atomic-result-list` component.
  */
@@ -34,11 +45,6 @@ export class AtomicResult {
   @Prop() content!: ParentNode;
 
   /**
-   * Whether this result should use `atomic-result-section-*` components.
-   */
-  @Prop() useSections = true;
-
-  /**
    * How results should be displayed.
    */
   @Prop() display: ResultDisplayLayout = 'list';
@@ -50,6 +56,8 @@ export class AtomicResult {
 
   /**
    * How large or small the visual section of results should be.
+   *
+   * This may be overwritten if an image size is defined in the result content.
    */
   @Prop() imageSize?: ResultDisplayImageSize;
 
@@ -65,13 +73,33 @@ export class AtomicResult {
     event.detail(this.result);
   }
 
+  private containsSections() {
+    return Array.from(this.content.children).some((child) =>
+      (resultSectionTags as readonly string[]).includes(
+        child.tagName.toLowerCase()
+      )
+    );
+  }
+
+  private getSection(section: typeof resultSectionTags[number]) {
+    return Array.from(this.content.children).find(
+      (element) => element.tagName.toLowerCase() === section
+    );
+  }
+
+  private getImageSizeFromSections() {
+    return (this.getSection('atomic-result-section-visual')?.getAttribute(
+      'image-size'
+    ) ?? undefined) as ResultDisplayImageSize | undefined;
+  }
+
   private getClasses() {
     const classes = getResultDisplayClasses(
       this.display,
       this.density,
-      this.imageSize ?? this.image
+      this.getImageSizeFromSections() ?? this.imageSize ?? this.image
     );
-    if (this.useSections) {
+    if (this.containsSections()) {
       classes.push('with-sections');
     }
     return classes;
